% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataScienceLab2.R
\name{drgplot}
\alias{drgplot}
\title{drgplot}
\usage{
drgplot(df, a)
}
\arguments{
\item{df}{a dataframe, use the dataset in the package drgdata2}

\item{a}{a string name for the payment variable in the datafram. Choose from Average Medicare payments, Average Total Payment, or Average Covered Charges}
}
\value{
drgplot returns a boxlpot of one of the payment variables by DRG
}
\description{
The function drgplot makes a boxplot for either the average Medicare payments, the average total payment, or the average covered charges by DRG code.
}
\examples{
drgplot(drgdata2, "Average Total Payments")

}
