% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataScienceLab2.R
\name{drgsummary}
\alias{drgsummary}
\title{drgsummary}
\usage{
drgsummary(df, a)
}
\arguments{
\item{df}{a dataframe, use the dataset in the package drgdata2}

\item{a}{a string name for the statistics, choose from mean, median, or sd for standard deviation}
}
\value{
drgsummary returns statistics of average Medicare payments
}
\description{
The function drgsummary calculates mean, median, or standard deviation for the average Medicare payments.
}
\examples{
drgsummary(drgdata2, "Median")

}
